Make `SomeJob::perform_(later|now)` type safe
Previously it was possible to enqueue all types of jobs with anything that
implemented `Serialize` and `Deserialize`. That wasn't very type safe
and runtime crashes waiting to happen.

This changes the `jobs!` macro such that you also have to supply the
type of argument your job expects. That way we can generate the code for
`::perform_later` and `::perform_now` to use that type for the argument.