Rewrite worker module and integration tests
I was not happy with the previous integration tests and how they
impacted the worker module.

The tricky bit is that we have to kill the workers somehow, otherwise
the tests will never finish. Sadly there has to be code to support that
in the worker module, unless we add `#[cfg(test)]` everywhere, but then
we don't get good test coverage and probably spaghetti code.

I'm still not completely satisfied with how the worker code is
structured, but I think it easier to follow now than it was before.

The tests also look much more like the code our users would write. I
think that is important so we don't loose the feeling of what our users
are writing.

This also changes the workers such that we always spawn one thread to
deal with the retry queue. So if you ask for 3 workers you'll actually
get 3 + 1. I think that is fine. One connection more to Redis shouldn't
be the end of the world.