Make connections generic over the type of queue
In the future I would like to implement different types of queues. I
don't think it makes sense to force users to always use Redis.

This required changing most of the core types because `WorkerConnection`
became `Connection<Q> where Q: JobQueue`. It was in particular
challenging with `Job` which needs to be object safe, but take the
connection as an argument.